Fix queue view resetting scroll position on swaps
Scroll still resets on track removal. This seems hard to fix.

The biggest change of this commit is actually the whole queue update
logic, which is now local-first, server-later. We stay in sync with the
server by carefully monitoring queue version changes and tracking the
expected server-side version as changes are applied locally.

This new logic is much more efficient both for the server and the
client. It also allows us to at least get rid of a near-full refresh of
the queue upon song swaps and removals. Also thanks to song swaps being
performed locally first (and not having to be locally realised later via
a full queue refresh) we have gotten rid of the jump-to-beginning bug
at least for swaps.

For some reason removals will still reset the queue location. Not even a
manual scroll position restoration in g_idle_add_once() helped.